requirements.txt
# Core libraries
playwright>=1.40.0
pydantic>=2

# LLM Provider SDKs (direct, no langchain wrapper)
anthropic>=0.25.0
openai>=1.3.0
groq>=0.4.0
azure-openai>=1.3.0

# System utilities
psutil>=5.9.0

# Utilities
python-dotenv>=1.0.0
typing-extensions>=4.5.0